overhaul consistency check
Now perform 3 checks. Given an edge (u, v) and an assignment (x, y),
all edges (a, b) adjacent to (u, v) should have either x, y or both in
their assigned edge. If we take the set of edges adjacent to u
there should be pair of edges such that one contains x but not
y and the other contains y but not x, since then they themselves
would not be adjacent.
Similarly give the set of edges adjacent to u and the edges adjacent
to v, there should be no pairs of edges with one edge from each set
such that both are adjacent to only x or only y since this would mean
both edges are connected through a single side of (u, v), and the
center edge has a dangling endpoint.