Edit loop test
test example wasn't bad. Since all the nodes were so close to the
correct path, the "bad" loop was still pretty close, and thus didn't
exemplify the extent of the possible problem.

Now it is clear that the optimal solution passes through the middle,
and the presence of loops throws the tree far from the optimal solution.