remove old matcher
Adding positive costs simply biases the solution towards shrinking.
This is a better bias than biasing towards growing with rewards since we
now do not create loops. However we would like to create indicators per
which would allow us to bias towards shrinking, but have a regularizer
make it difficult to shrink too far away from the true node location.

Implementing the node constraints on our old formulation was
non trivial. Implementing them would require adding node-node match
indicators, Along with the appropriate constraints to make sure nodes
are matched properly. Next we need to tie the node indicators to edge
indicators, so we would also need to implement the constraints on
matched nodes must also match "in" edges and "out" edges. At this point
we would have implemented all of the constraints from the new approach,
except the balance constraint and the degree constraint.

Thus, we can see if it would be worth keeping the old transition
constraints. If we kept them, we have `a * (n choose k)` indicators per node,
(where a is the number of possible node assignments.
n is the number of possible adjacent edge assignments.
and k is the degree of the assigned node in the Tree.)
of indicators per node, and 1 constraint per indicator.
Without them we have no extra indicators and 1 constraint per node
and possible adjacent edge for balance,
assignment. plus 1 constraint per node for the degree.

It does not make sense to keep the old version.