Add a nonempty_placeholder term
Since the random location node picks a random point, and then
computes the random shift such that the chosen point is in the center
of the roi, it doesn't matter if you shift your desired roi before the
random location, it will not provide an equivalent shift in the roi
requested from random location.

To solve this, I simply request a placeholder of the original size for
which the node will be centered, providing the exact same random
shift as before. However now my desired roi for the non-placeholder
arrays are offset within the larger array, thus we get the final roi as
desired