Tree seperation
Overhauled the code for guaranteeing a minimum distance between
two graphs. It was originally done by calculating a distance transform
and then allowing O(1) look up for distance per node in the comparison
graph. This still ends up being slow
however since generating the distance transform is slow, and we still
need to iterate over all the points in one graph.
It turns out putting a graph in a cKDTree allows for the quick
retrieval of all pairs of nodes that are too close, and then finer
detail when it comes to actually calculating the distances, and
optimal move direction.